requirements.txt
banidb
fastapi 
uvicorn 
scikit-learn 
pandas 
numpy 
motor 
nltk 
indic-nlp-library